Changed float to double.

+++ main.cpp
@@ -41,9 +41,9 @@ std::unique_ptr<SDL_Texture, decltype(tex_deleter)> mktxt(const char* txt);
 
 // x,y are relative screen coordinates, 0 being top/left and 1 being bottom/right.
 // w,h are screen resolution.
-void rendermid(SDL_Texture* tex, float x, float y, int w, int h);
+void rendermid(SDL_Texture* tex, double x, double y, int w, int h);
 
-float lerp(float t, float x0, float x1, float t0, float t1) { return x0 + (t - t0) / (t1 - t0) * (x1 - x0); };
+double lerp(double t, double x0, double x1, double t0, double t1) { return x0 + (t - t0) / (t1 - t0) * (x1 - x0); };
 
 int main(int argc, char **argv)
 {
@@ -158,11 +158,11 @@ int main(int argc, char **argv)
         }
 
         // The current position of Mr.Point (in relative screen-coordinates).
-        float x = 0.01f, y = 0.01f;
+        double x = 0.01, y = 0.01;
 
         // Update the dot's position according to the "storyline".
         if (state == STATE_RECORDING) {
-            float t = 0.001f*(SDL_GetTicks() - t0);
+            double t = 0.001*(SDL_GetTicks() - t0);
 
             // That's the choreography!
             if (0 <= t && t < 3) {
@@ -305,7 +305,7 @@ std::unique_ptr<SDL_Texture, decltype(tex_deleter)> mktxt(const char* txt)
     return tex;
 };
 
-void rendermid(SDL_Texture* tex, float x, float y, int w, int h)
+void rendermid(SDL_Texture* tex, double x, double y, int w, int h)
 {
     // Get the texture w/h.
     int tw, th;
@@ -313,8 +313,8 @@ void rendermid(SDL_Texture* tex, float x, float y, int w, int h)
 
     //Setup the destination rectangle to be at the (pixel) position we want.
     SDL_Rect dst;
-    dst.x = int(x*w - tw*0.5f);
-    dst.y = int(y*h - th*0.5f);
+    dst.x = int(x*w - tw*0.5);
+    dst.y = int(y*h - th*0.5);
 
     //Query the texture to get its width and height to use
     SDL_QueryTexture(tex, NULL, NULL, &dst.w, &dst.h);
